+++ ed25519/Cargo.toml
@@ -22,8 +22,7 @@ signature = { version = "3", default-features = false }
 
 # optional dependencies
 pkcs8 = { version = "0.11", optional = true }
-serde = { version = "1", optional = true, default-features = false }
-serde_bytes = { version = "0.11", optional = true, default-features = false }
+serdect = { version = "0.4", optional = true, default-features = false }
 zeroize = { version = "1", optional = true, default-features = false }
 zerocopy = { version = "0.8", optional = true, features = ["derive"] }
 
@@ -37,7 +36,7 @@ rand_core = { version = "0.9", features = ["std"] }
 default = ["alloc"]
 alloc = ["pkcs8?/alloc"]
 pem = ["alloc", "pkcs8/pem"]
-serde_bytes = ["serde", "dep:serde_bytes"]
+serde = ["dep:serdect"]
 
 [lints]
 workspace = true
